requirements.txt
polars~=0.20
pandas~=2.2
scikit-learn~=1.4
statsmodels~=0.14
mlxtend~=0.23
ollama~=0.1
pyyaml~=6.0
playwright~=1.42
psutil~=5.9
